refactor/SW-45: 컨벤션에 따라 북마크 삭제 로직 리팩터링
- Controller 전용 DTO(BookmarkRequests) 도입
- 커스텀에러 적용

+++ src/main/java/com/web/SearchWeb/bookmark/service/BookmarkService.java
@@ -25,7 +25,7 @@ public interface BookmarkService {
                        String note, Long primaryCategoryId, String tags);
 
     //북마크 삭제
-    int deleteBookmark(Long memberId, Long bookmarkId);
+    Long deleteBookmark(Long memberId, Long bookmarkId);
 
     // 링크 조회 또는 생성 (URL 정규화)
     Link getOrCreateLink(String url, Long createdByMemberId);
